Move case expression tests into new file
